[README] - add screenshots

+++ src/vector/vector.h
@@ -5,8 +5,8 @@
 
 struct vector2
 {
-  int x;
-  int y;
+  double x;
+  double y;
 };
 
 struct vector2 *create_vect(void);
@@ -19,6 +19,6 @@ struct vector2 *mult_two_vects(struct vector2 a, struct vector2 b);
 
 struct vector2 *mult_vect_coef(struct vector2 a, float coef);
 
-int distance_vect(struct vector2 a, struct vector2 b);
+double distance_vect(struct vector2 a, struct vector2 b);
 
 #endif /* !VECTOR_H */
